move gptq patching to export stage
